send to core
* send read_evl and read_evr to core.py from utils.py for easier reading
